requirements.txt
Django>=5.0.2
pillow==10.2.0
gunicorn==21.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.1
django-cors-headers==4.1.0
dj-rest-auth[with_social]==5.0.2
whitenoise==6.6.0
django-allauth~=0.57.0
